// src/features/mapping/types.ts
// ─── Mapping ─────────────────────────────────────────────────────────────────

export interface CpCoin {
  id: string;
  symbol: string;
  name: string;
}

export interface GeckoCoin {
  id: string;
  symbol: string;
  name: string;
}